Refactor screens and add intro animation
