+++ Desktop/AI_HACK/koa-coach/app/(home)/therapist.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import { Buffer } from "buffer";
 import {
   View,
@@ -11,12 +11,21 @@ import {
   Image,
   Animated,
   Easing,
+  useWindowDimensions,
 } from "react-native";
+import Video from "react-native-video";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import Anthropic from "@anthropic-ai/sdk";
 import { Audio } from "expo-av";
 import * as FileSystem from "expo-file-system";
 import { FontAwesome, MaterialIcons } from "@expo/vector-icons";
+import anim1 from "../animations/anime1.mp4";
+import anim2 from "../animations/anime2.mp4";
+import anim3 from "../animations/anime3.mp4";
+import anim4 from "../animations/anime4.mp4";
+import anim5 from "../animations/anime5.mp4";
+import anim6 from "../animations/anime6.mp4";
+import avatarImage from "../animations/koa.png";
 
 const flagIcons = {
   en: require("../../assets/flags/uk.png"),
@@ -46,6 +55,14 @@ const TherapistChat = () => {
     pronouns: "",
   });
 
+  const { width, height } = useWindowDimensions();
+  const avatarSize = Math.min(width, height) * 0.7;
+
+  const [isAnimating, setAnimating] = useState(false);
+  const [currentVideo, setCurrentVideo] = useState(null);
+  const videoIndex = useRef(0);
+  const videoRef = useRef(null);
+
   const pulseAnim = useState(new Animated.Value(1))[0];
 
   useEffect(() => {
@@ -70,6 +87,29 @@ const TherapistChat = () => {
       pulseAnim.setValue(1);
     }
   }, [isRecording]);
+
+  const animations = [anim1, anim2, anim3, anim4, anim5, anim6];
+
+  const playNextAnimation = () => {
+    const next = animations[videoIndex.current % animations.length];
+    videoIndex.current = (videoIndex.current + 1) % animations.length;
+    setCurrentVideo(next);
+    setAnimating(true);
+  };
+
+  const handleClick = () => {
+    if (isAnimating) {
+      setAnimating(false);
+      setCurrentVideo(null);
+    } else {
+      playNextAnimation();
+    }
+  };
+
+  const handleVideoEnd = () => {
+    setAnimating(false);
+    setCurrentVideo(null);
+  };
 
   const anthropic = new Anthropic({
     apiKey: process.env.EXPO_PUBLIC_ANTHROPIC_API_KEY,
@@ -344,6 +384,29 @@ const TherapistChat = () => {
         ))}
       </View>
 
+      <View style={[styles.avatarWrapper, { width: avatarSize, height: avatarSize }]}>
+        <TouchableOpacity onPress={handleClick}>
+          {isAnimating && currentVideo ? (
+            <Video
+              key={currentVideo}
+              ref={videoRef}
+              source={currentVideo}
+              resizeMode="contain"
+              repeat={false}
+              paused={false}
+              onEnd={handleVideoEnd}
+              style={[styles.avatar, { width: avatarSize, height: avatarSize }]}
+            />
+          ) : (
+            <Image
+              source={avatarImage}
+              style={[styles.avatar, { width: avatarSize, height: avatarSize }]}
+              resizeMode="contain"
+            />
+          )}
+        </TouchableOpacity>
+      </View>
+
       <FlatList
         data={messages}
         renderItem={renderMessage}
@@ -436,6 +499,16 @@ const styles = StyleSheet.create({
     color: "#fff",
     fontSize: 16,
     fontWeight: "600",
+  },
+  avatarWrapper: {
+    alignItems: "center",
+    justifyContent: "center",
+    marginVertical: 20,
+    backgroundColor: "#000",
+    overflow: "hidden",
+  },
+  avatar: {
+    backgroundColor: "#000",
   },
 });
